requirements.txt
dj-static==0.0.6
Django==1.8.2
django-braces==1.8.0
Pillow==2.8.2
reportlab==3.2.0
six==1.9.0
static3==0.6.1
wheel==0.24.0
django-toolbelt==0.0.1
wsgiref==0.1.2
dj-database-url==0.3.0
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
pylibmc==1.4.0
django-pylibmc==0.5.0